requirements.txt
anyio==4.9.0
asttokens==3.0.0
babel==2.17.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
decorator==5.2.1
executing==2.2.0
Flask==3.1.1
h11==0.16.0
httptools==0.6.4
idna==3.10
ipdb==0.13.13
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
parso==0.8.4
pillow==12.0.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scapy==2.6.1
sniffio==1.3.1
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.5.0
uvicorn==0.23.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
PyNaCl==1.5.0
